# docs/swagger.yaml
basePath: /
definitions:
...
  requests.StudentPayload:
    properties:
      classroom_id:
        type: integer
      name:
        type: string
    type: object
  requests.StudentUpdatePayload:
    properties:
      classroom_id:
        type: integer
      name:
        type: string
    type: object
...
  resources.ClassroomResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  resources.StudentCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/resources.StudentResource'
        type: array
      meta:
        $ref: '#/definitions/resources.PaginationMeta'
    type: object
  resources.StudentResource:
    properties:
      classroom:
        $ref: '#/definitions/resources.ClassroomResource'
      classroom_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.ErrorBody:
    properties:
      details: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  title: Safa API
  version: "1.0"
paths:
...
  /students:
    get:
      description: Get a paginated list of Student records with optional filters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 15)
        in: query
        name: per_page
        type: integer
      - description: Search by name
        in: query
        name: search
        type: string
      - description: Filter by classroom ID
        in: query
        name: classroom_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.StudentCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: List all Student records
      tags:
      - Student
    post:
      consumes:
      - application/json
      parameters:
      - description: Student payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.StudentPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.StudentResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a Student
      tags:
      - Student
  /students/{id}:
    delete:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a Student
      tags:
      - Student
    get:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.StudentResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a Student by ID
      tags:
      - Student
    put:
      consumes:
      - application/json
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.StudentUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.StudentResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a Student
      tags:
      - Student
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"